Configuration to enable Table of Contents properly
* params.page.toc enables the theme's toc. Maintains a full list of
heading links with indents on the side, under the Contents panel.
Contents moves with scroll and highlighting shows current position among
toc links. auto as false means the menu doesn't hide links that are not
close to reading point.
* markup.tableOfContents controls if the list is numbered and also what
heading levels it supports. Seems like a global rule of thumb is for h1
to be reserved for the article title and the anchors should start from
h2 onwards, hence startLeve =2 and endLevel =6.